send more to delete endpoint

+++ resources/js/filament-uppy-component.js
@@ -120,20 +120,26 @@ window.fileUploaderComponent = function fileUploaderComponent({
         },
 
         removeCompletedFile(index) {
-            const key = this.state[index].url.split('/').pop();
-            const uuid = key.split('.')[0];
+            const file = this.state[index];
 
-            this.state.splice(index, 1);
+            if (!!file) {
+                this.state.splice(index, 1);
 
-            if (!!deleteEndpoint) {
-                fetch(deleteEndpoint, {
-                    method: 'DELETE',
-                    headers: {
-                        'Content-Type': 'application/json',
-                        'X-CSRF-TOKEN': document.querySelector('meta[name="csrf-token"]').content,
-                    },
-                    body: JSON.stringify({ uuid }),
-                });
+                if (!!deleteEndpoint) {
+                    const key = file.url.split('/').pop();
+                    const uuid = key.split('.')[0];
+                    const name = file.name;
+                    const url = file.url;
+
+                    fetch(deleteEndpoint, {
+                        method: 'DELETE',
+                        headers: {
+                            'Content-Type': 'application/json',
+                            'X-CSRF-TOKEN': document.querySelector('meta[name="csrf-token"]').content,
+                        },
+                        body: JSON.stringify({ name, url, uuid }),
+                    });
+                }
             }
         },
